pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = ""
authors = [
    {name = "Jaakko Airikkala",email = "jaakkel@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "google-generativeai (>=0.8.3,<0.9.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
